Engine Library Changes to Static Library, Warnings Removed

+++ ProjectGE/src/ProjectGE/ImGui/ImGuiLayer.cpp
@@ -77,7 +77,7 @@ namespace ProjectGE {
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
 		Window& window = app.GetWindow();
-		io.DisplaySize = ImVec2(window.GetWidth(), window.GetHeight());
+		io.DisplaySize = ImVec2((float)window.GetWidth(), (float)window.GetHeight());
 		ImGui::Render();
 
 		Renderer* renderer = window.GetRenderer();
